requirements.test.txt
pyserial==3.5
pytest
pytest-homeassistant-custom-component==0.13.107
async-timeout==4.0.3